Refactor assigned shifts section to AssignedShifts component
