+++ web/src/components/RequestStatusCard/RequestStatusCard.tsx
@@ -1,6 +1,5 @@
-import { useMemo, useState } from 'react'
+import { useState } from 'react'
 
-import { PopoverContent } from '@radix-ui/react-popover'
 import { format } from 'date-fns'
 import { nl } from 'date-fns/locale/nl'
 import {
@@ -8,10 +7,8 @@ import {
   CalendarDays,
   Clock,
   Edit,
-  Mail,
   MapPin,
   MousePointerClick,
-  Phone,
   Users,
 } from 'lucide-react'
 import { WorkRequestsQuery } from 'types/graphql'
@@ -20,11 +17,11 @@ import { Link, routes } from '@redwoodjs/router'
 
 import { formatAddress } from 'src/lib/formatAddress'
 
+import AssignedShifts from '../AssignedShifts/AssignedShifts'
 import PlanWorkComponent from '../PlanWorkComponent/PlanWorkComponent'
 import { Badge } from '../ui/badge'
 import { Button } from '../ui/button'
 import { Card, CardContent, CardHeader, CardTitle } from '../ui/card'
-import { Popover, PopoverTrigger } from '../ui/popover'
 import { Progress } from '../ui/progress'
 import { Separator } from '../ui/separator'
 
@@ -34,24 +31,6 @@ type RequestStatusCardProps = {
 }
 const RequestStatusCard = ({ className, request }: RequestStatusCardProps) => {
   const [editOpen, setEditOpen] = useState(false)
-
-  const agenciesWithShiftCounts = useMemo(() => {
-    if (request?.shifts == null) return
-    return request.shifts.reduce((acc, current) => {
-      const currentAgencyId = current.tempAgency.id
-      const existingAgency = acc.find((agency) => agency.id === currentAgencyId)
-
-      if (existingAgency) {
-        existingAgency.count++
-      } else {
-        acc.push({
-          ...current.tempAgency,
-          count: 1,
-        })
-      }
-      return acc
-    }, [])
-  }, [request])
 
   return (
     <Card className={className}>
@@ -168,55 +147,7 @@ const RequestStatusCard = ({ className, request }: RequestStatusCardProps) => {
             </div>
           </>
         )}
-        {request.status === 'CONFIRMED' && (
-          <>
-            <Separator className="opacity-40" />
-            <div className="flex items-center gap-1">
-              <Users />
-              <h3>Toegewezen medewerkers</h3>
-            </div>
-            <div className="ml-10 flex flex-col gap-2">
-              {agenciesWithShiftCounts.map((agency) => {
-                return (
-                  <div
-                    key={agency.id}
-                    className="flex items-center gap-5 font-semibold"
-                  >
-                    {request.numWorkers}{' '}
-                    <span className="text-sm font-normal">van</span>
-                    <Popover>
-                      <PopoverTrigger className="hover:underline">
-                        {agency.name}
-                      </PopoverTrigger>
-                      <PopoverContent
-                        side="top"
-                        className="my-1 min-w-fit rounded-md bg-gray-900 px-4 py-2 text-sm text-primary-foreground"
-                      >
-                        <div className="flex flex-col gap-1">
-                          <h3 className="mb-1 text-accent">{agency.name}</h3>
-                          <div className="flex items-center gap-1">
-                            <MapPin className="size-4" />
-                            {formatAddress(agency.address)}
-                          </div>
-                          <div className="flex items-center gap-1">
-                            <Phone className="size-4" />
-                            <a href={`tel:${agency.phone}`}>{agency.phone}</a>
-                          </div>
-                          <div className="flex items-center gap-1">
-                            <Mail className="size-4" />
-                            <a href={`mailto:${agency.email}`}>
-                              {agency.email}
-                            </a>
-                          </div>
-                        </div>
-                      </PopoverContent>
-                    </Popover>
-                  </div>
-                )
-              })}
-            </div>
-          </>
-        )}
+        {request.status === 'CONFIRMED' && <AssignedShifts request={request} />}
         {request.status === 'DONE' && (
           <>
             <Separator className="opacity-40" />
